# public/openapi.yaml
openapi: 3.0.3
info:
  title: CoreGRE API Documentation
  description: |
    API REST per il sistema ERP CoreGRE - Gestione calzaturiera

    ## Autenticazione
    Le API mobile utilizzano autenticazione tramite username/password (PIN).
    Gli operatori sono gestiti nella tabella `inwork_operators` e devono avere status `active = 1` per autenticarsi.
    Ogni operatore ha permessi granulari per accedere ai moduli (quality, repairs) tramite la tabella `inwork_module_permissions`.

    ## Versioning
    Versione corrente: 1.0.0

    ## Rate Limiting
    Nessun rate limiting attualmente implementato.

  version: 1.0.0
  contact:
    name: Stefano Solidoro
    email: kishosa@me.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:84
    description: Server di sviluppo locale
  - url: https://localhost:84
    description: Server di produzione
...
paths:
...
  /api/mobile/profile:
    get:
      tags:
        - Mobile API
      summary: Ottiene profilo operatore
      description: Recupera profilo completo e statistiche operatore
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
          description: ID operatore
        - name: X-App-Type
          in: header
          schema:
            type: string
            enum: [quality, repairs]
            default: quality
      responses:
        '200':
          description: Profilo recuperato
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      user:
                        type: string
                      full_name:
                        type: string
                      reparto:
                        type: string
                      statistiche:
                        type: object
                      app_type:
                        type: string
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/mobile/daily-summary:
    get:
      tags:
        - Mobile API
      summary: Riepilogo giornaliero operatore
      description: Recupera riepilogo attività giornaliera per operatore
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
          description: ID operatore
        - name: data
          in: query
          schema:
            type: string
            format: date
            example: "2025-01-15"
          description: Data (default oggi)
        - name: X-App-Type
          in: header
          schema:
            type: string
            enum: [quality, repairs]
      responses:
        '200':
          description: Riepilogo recuperato
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      operatore:
                        type: object
                      data:
                        type: string
                      app_type:
                        type: string
                      summary:
                        type: object
  /api/mobile/system-data:
    get:
      tags:
        - Mobile API
      summary: Dati di sistema comuni
      description: |
        Recupera dati di sistema (reparti, linee, taglie, ecc.)
        Supporta filtri per tipo di dati richiesti
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, reparti, linee, taglie, quality, repairs]
            default: all
        - name: nu
          in: query
          schema:
            type: string
          description: ID numerata (richiesto per type=taglie)
      responses:
        '200':
          description: Dati recuperati
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      reparti:
                        type: array
                        items:
                          type: object
                      linee:
                        type: array
                        items:
                          type: object
                      taglie:
                        type: object
                      reparti_hermes:
                        type: array
                        items:
                          type: object
                      difetti:
                        type: array
                        items:
                          type: object
                      laboratori:
                        type: array
                        items:
                          type: object
                      causali_frequenti:
                        type: array
                        items:
                          type: string
...
  /api/mobile/enabled-modules:
    get:
      tags:
        - Mobile API
      summary: Moduli abilitati per operatore
      description: |
        Recupera i moduli abilitati per un operatore specifico.
        Ritorna sia la lista dei moduli abilitati che tutti i moduli con il loro stato.
      parameters:
        - name: operator_id
          in: query
          required: true
          schema:
            type: integer
          description: ID dell'operatore
      responses:
        '200':
          description: Moduli recuperati
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Moduli recuperati con successo
                  data:
                    type: object
                    properties:
                      operator_id:
                        type: integer
                        example: 1
                      enabled_modules:
                        type: array
                        description: Lista moduli abilitati
                        items:
                          type: string
                          enum: [quality, repairs]
                        example: ["quality", "repairs"]
                      all_modules:
                        type: object
                        description: Tutti i moduli con stato enabled
                        properties:
                          quality:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "Controllo Qualità"
                              enabled:
                                type: boolean
                                example: true
                          repairs:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "Riparazioni Interne"
                              enabled:
                                type: boolean
                                example: true
        '404':
          description: Operatore non trovato o non attivo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Operatore non trovato o non attivo

  # ==================== QUALITY API ====================
...
  /api/quality/operator-daily-summary:
    get:
      tags:
        - Quality API
      summary: Riepilogo giornaliero operatore CQ
      description: Statistiche e lista controlli giornalieri per operatore
      parameters:
        - name: operatore
          in: query
          required: true
          schema:
            type: string
        - name: data
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Riepilogo recuperato
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      data:
                        type: string
                      total_controls:
                        type: integer
                      total_exceptions:
                        type: integer
                      controls_list:
                        type: array
                        items:
                          type: object
                      top_defects:
                        type: array
                        items:
                          type: object
                      top_departments:
                        type: array
                        items:
                          type: object
  /api/quality/record-details:
    get:
      tags:
        - Quality API
      summary: Dettagli record controllo
      description: Recupera dettagli completi di un record di controllo qualità
      parameters:
        - name: record_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dettagli recuperati
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      record:
                        type: object
                      exceptions:
                        type: array
                        items:
                          type: object
...
  /api/riparazioni-interne:
    get:
      tags:
        - Riparazioni API
      summary: Lista riparazioni
      description: Recupera lista riparazioni con filtri opzionali
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [complete, incomplete, all]
        - name: operatore
          in: query
          schema:
            type: string
        - name: data
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista riparazioni
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InternalRepair'
    post:
      tags:
        - Riparazioni API
      summary: Crea nuova riparazione
      description: Crea un nuovo record di riparazione interna
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalRepairCreate'
      responses:
        '201':
          description: Riparazione creata
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
  /api/riparazioni-interne/show:
    get:
      tags:
        - Riparazioni API
      summary: Dettagli riparazione
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dettagli riparazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/InternalRepair'
...
  /api/riparazioni-interne/stats:
    get:
      tags:
        - Riparazioni API
      summary: Statistiche riparazioni
      parameters:
        - name: operatore
          in: query
          schema:
            type: string
        - name: periodo
          in: query
          schema:
            type: string
            enum: [today, week, month]
      responses:
        '200':
          description: Statistiche recuperate

  # ==================== DASHBOARD API ====================
...
components:
  schemas:
...
    InternalRepair:
      type: object
      properties:
        ID:
          type: integer
        CARTELLINO:
          type: string
        COMMESSA:
          type: string
        CAUSALE:
          type: string
        OPERATORE:
          type: string
        REPARTO:
          type: string
        LABORATORIO:
          type: integer
        DATA:
          type: string
        COMPLETA:
          type: integer
          enum: [0, 1]
        P01:
          type: integer
        P02:
          type: integer
        # ... P03-P20
        NOTE:
          type: string
    InternalRepairCreate:
      type: object
      properties:
        cartellino:
          type: string
        commessa:
          type: string
        causale:
          type: string
        operatore:
          type: string
        reparto:
          type: string
        laboratorio:
          type: integer
        data:
          type: string
          format: date
        taglie:
          type: object
          description: Oggetto con P01-P20 e quantità
        note:
          type: string
      required:
        - cartellino
        - operatore
        - reparto
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Si è verificato un errore
        code:
          type: string
          example: ERR_001
  responses:
...
    NotFoundError:
      description: Risorsa non trovata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key per autenticazione (futuro)

# Security applicata globalmente (attualmente non implementata)
# security:
#   - ApiKeyAuth: []